% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_plot.R
\name{anova_plot}
\alias{anova_plot}
\title{One-Way Anova Plotting}
\usage{
anova_plot(data, x, y)
}
\arguments{
\item{data}{a data frame containing the x & y variables}

\item{x}{independent categorical variable}

\item{y}{dependent continuous variable}
}
\value{
a list with a ggplot2 plot object named plot and a anova summary table in a tableGrob
object named anova
}
\description{
Generates the anova summary table and the box plot for specified data in an easy to plot form
}
\examples{
library(gridExtra)
x <- anova_plot(iris, x = Species, y = Petal.Length)
marrangeGrob(x, ncol = 2, nrow = 2)

}
