% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_cases.R
\name{complete_cases}
\alias{complete_cases}
\title{Removes NA Observations By Column}
\usage{
complete_cases(data, ...)
}
\arguments{
\item{data}{a data frame}

\item{...}{columns to remove NA obs, defaults to all arguments; can also apply dplyr special functions}
}
\value{
a data frame with duplicated rows by given columns
}
\description{
Removes NA Observations By Column
}
\examples{
d <- data.frame(x1 = rep(1:3, each=4), x2 = rep(1:4, each=3), z = rep(1:2, 6), a = rep(1:6, 2))
complete_cases(data = d, x2, z)
complete_cases(d, a, z)
complete_cases(d)
complete_cases(d, starts_with("x"))

}
\seealso{
\code{\link{remove_duplicated}} to remove duplicates
}
