% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{diagnostic_plots}
\alias{diagnostic_plots}
\title{Diagnostic Plots for lm and glm}
\usage{
diagnostic_plots(mod)
}
\arguments{
\item{mod}{a lm or glm model obj}
}
\value{
a list of diagnostic plots
}
\description{
Generates the diagnostic plots in ggplot2, similar to the ones that are generated via the plot command. Currently available for lm and glm models
}
