% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/duplicated_data.R
\name{duplicated_data}
\alias{duplicated_data}
\title{Show Duplicated Rows}
\usage{
duplicated_data(data, ...)
}
\arguments{
\item{data}{a data frame}

\item{...}{columns to search for duplicated values; defaults to all arguments}
}
\value{
a data table with duplicated rows by given columns
}
\description{
Finds all duplicated rows by given columns. Equivalent to doing data[duplicated(data) | duplicated(data, fromLast = TRUE), ]
}
\examples{
d <- data.frame(x = rep(1:3, each=4), y = rep(1:4, each=3), z = rep(1:2, 6), a = rep(1:6, 2))
duplicated_data(data = d, y, z)
duplicated_data(d, a, z)
duplicated_data(d)
}

