% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_iterative_list.R
\name{is_iterative_list}
\alias{is_iterative_list}
\title{Iterative List Check}
\usage{
is_iterative_list(l)
}
\arguments{
\item{l}{a list}
}
\value{
logical; is l an iterative list?
}
\description{
An iterative list is one which is generated from an lapply loop.
The lapply function call should return a list with consistent names.
Therefore each element within l should be a list with the same names.
No element in l should be null, but it is ok if an element of l
consists of a named items that are null
}
