% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logging}
\alias{logging}
\alias{start_logging}
\alias{get_logfile}
\alias{logdf}
\alias{stopQuietly}
\alias{logerror}
\title{Logging Functions}
\usage{
start_logging(name = paste("R", str_replace(Sys.time(), " ", "_"), sep = "_"))

get_logfile()

logdf(df)

stopQuietly()

logerror(msg)
}
\description{
Additional functions to be used with the logging package
}
\details{
start_logging() to start a new logging session \cr

get_logfile() to obtain the name of the logfile \cr

logdf() to log a data.frame \cr

stopQuietly() quits without an error message \cr

logerror() logs an error and quits the program \cr
}
