% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_outliers_na.R
\name{make_outliers_na}
\alias{make_outliers_na}
\title{Makes Outliers NA}
\usage{
make_outliers_na(x, multiplier)
}
\arguments{
\item{x}{a numeric vector, with more than 2 values}

\item{multiplier}{a number greater than 0, sd is multiplied by this to determine boundaries}
}
\value{
numeric vector with outliers NA
}
\description{
Makes all values beyond specified standard deviations from the mean NA.
}
\examples{
make_outliers_na(rnorm(1000), 2)

}
