% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile_bin.R
\name{percentile_bin}
\alias{percentile_bin}
\title{Group Vector by its Percentiles}
\usage{
percentile_bin(x, opt_n_bins = 5, group_prefix = "group ")
}
\arguments{
\item{x}{a numeric vector}

\item{opt_n_bins}{int, number of bins/groups to generate. Percentile splits are equally spaced}

\item{group_prefix}{string, the prefix to add to the bin labels}
}
\value{
a vector with opt_n_bins factors
}
\description{
Given the number of equally spaced bins, groups a vector by its percentiles
}
\examples{
x <- rpois(20, 4)
percentile_bin(x)
}

