% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdata.R
\name{rdata}
\alias{rdata}
\title{Random Data Frame with Specified Column Types}
\usage{
rdata(n, ..., .id = TRUE)
}
\arguments{
\item{n}{number of rows for the new data frame}

\item{...}{column specifications; use rdata helpers: \link{generate}}

\item{.id}{logical; if TRUE, generate a unique numeric id for each row; else no id}
}
\value{
A data frame
}
\description{
Generate a random data frame with a specified number of rows and randomly generated columns.
Column specifications are generated via \code{\link{generate}} functions.
}
\examples{
rdata(n = 50, gen_random(ncol = 2))

# generate lists
ex1 <- gen_num(name = "x", values = 1:20, replace = FALSE, add.na = TRUE)
ex2 <- gen_char(3, letters, c("a", "b", "c"))
ex3 <- gen_bool(name = "true_or_false", probs = c(.75, .25))
state1 <- gen_state(kind = "abb")
state2 <- gen_state(kind = "full")
gender <- gen_gender(name = "sex")

# generate random data frame
rdata(n = 100, ex1, ex2, ex3, gender, state1, state2, .id = FALSE)

}
