% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/refine.R
\name{refine}
\alias{refine}
\alias{rm_na_list}
\title{Refine}
\usage{
refine(obj, f)

rm_na_list(list)
}
\arguments{
\item{obj}{An object to be filtered}

\item{f}{Function to apply on each element of object. Function should return a logical vector}
}
\value{
a object that has been filtered as requested
}
\description{
Wrapper for Filter. (change the order of args)
Returns subset of lists that meets conditions supplied by a function.
Applies the function f onto the list and returns that which are true.
}
\examples{
l <- list(NULL, mtcars, iris, NULL, mtcars, NULL, iris)
refine(l, function(x) !is.null(x))
refine(l, function(x) x \%>\% colnames \%>\% str_detect("mpg") \%>\% any)
rm_na_list(l)
}

