% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_list.R
\name{rename_list}
\alias{rename_list}
\title{Rename Data Nested within a List}
\usage{
rename_list(my_list, names = NULL)
}
\arguments{
\item{my_list}{a list; list elements may or may not have names}

\item{names}{character vector; the names of the elements of the list, if the list is unnamed}
}
\value{
a list
}
\description{
Loops through the elements of a list and adds a name_id column to data.frames 
corresponding to the provided names.
}
\examples{

#############
# Example 1 #
#############
paths <- list.files(directory)
files <- lapply(paths, fread)
rename_list(list = files, names = paths)

#############
# Example 2 #
#############
l <- purrr::rerun(5, x = list(a = head(mtcars), b = head(iris), c = list(o = 1, p = list(q = head(airquality), u = 1:10))), y = head(mtcars))
rename_list(l, names = paste0("id", 1:5))

names(l) <- paste0("id", 1:5)
rename_list(l)
}
