% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reorder_cols.R
\name{reorder_cols}
\alias{reorder_cols}
\title{Reorder Columns in Data Frame}
\usage{
reorder_cols(df, ...)
}
\arguments{
\item{df}{a data frame}

\item{...}{arguments to reorder; arguments should be in the order they
should appear on the ordered data frame}
}
\value{
a data frame whose columns are ordered as specified
}
\description{
Brins selected columns at the beginning and adds other columns at the end.
Compatible with dplyr's special select functions.
}
\examples{
df <- plyr::rdply(5, rnorm(26))
colnames(df) <- c("id", letters)
reorder_cols(df, matches("[aeiou]"))
}

