% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdata_helpers_shortcuts.R
\name{shortcut_generators}
\alias{gen_age}
\alias{gen_directions}
\alias{gen_gender}
\alias{gen_month}
\alias{gen_race}
\alias{gen_rank}
\alias{gen_state}
\alias{gen_subject}
\alias{gen_year}
\alias{shortcut_generators}
\title{rdata Argument Generators}
\usage{
gen_gender(ncol = 1, name = "gender", probs = NULL, add.na = FALSE,
  as.factor = TRUE)

gen_age(ncol = 1, name = "age", replace = TRUE, add.na = FALSE,
  as.factor = FALSE)

gen_race(ncol = 1, name = "race", replace = TRUE, probs = NULL,
  add.na = FALSE, as.factor = TRUE)

gen_directions(ncol = 1, name = "directions", probs = NULL,
  add.na = FALSE, as.factor = TRUE)

gen_state(ncol = 1, name = "state", kind = "full", probs = NULL,
  add.na = FALSE, as.factor = TRUE)

gen_rank(ncol = 1, name = "rank", probs = NULL, add.na = FALSE,
  as.factor = TRUE)

gen_subject(ncol = 1, name = "subject", probs = NULL, add.na = FALSE,
  as.factor = TRUE)

gen_month(ncol = 1, name = "month", kind = "full", probs = NULL,
  add.na = FALSE, as.factor = TRUE)

gen_year(start = 1500, end = format(Sys.Date(), "\%Y"), ncol = 1,
  name = "year", replace = TRUE, add.na = FALSE, as.factor = FALSE)
}
\value{
an rdata_generator list object that can be passed into \code{rdata} to generate a columns of the random data frame
}
\description{
Shortcut rdata generators. Values are predefined.
}
\details{
Default values: \cr

gen_gender: male, female \cr
gen_age: integers from 0 to 100 \cr
gen_race: white, black, hispanic, asian, native, multi \cr
gen_directions: north, west, south, east \cr
gen_state: all US states; for full name set kind = "full", for abbreviations set kind = "abb" \cr
gen_rank: low, medium, high \cr
gen_subject: math, science, read, ss \cr
gen_month: months of the year; for full name set kind = "full", for abbreviations set kind = "abb" \cr
gen_year: sequence of integers from 1500 to current year; start and end year are customizable \cr
}

