% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str_scan.R
\name{str_scan}
\alias{str_scan}
\title{Split and Scan a String}
\usage{
str_scan(string, index = 1, delimiter = " ")
}
\arguments{
\item{string}{input vector}

\item{index}{position of string want to isolate after splitting}

\item{delimiter}{what to split the string by}
}
\value{
string vector
}
\description{
Splits the string by the delimiter and grabs the index-th value
}
\examples{
x <- c("language1_r", "language2_python", "language3_java")
str_scan(x, 2, "_")
}

