% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_plot.R
\name{table_plot}
\alias{table_plot}
\title{Plots Data Frames}
\usage{
table_plot(df, title = "", rownames = NULL)
}
\arguments{
\item{df}{a data frame to be turned into a plot object}

\item{title}{title of the plot object}

\item{rownames}{rownames of the data frame; defaults to original rownames of data frame}
}
\value{
a grob with table that can be plotted
}
\description{
Generates a table grob object for plotting with ggplot2.
}
\examples{
table_plot(head(mtcars), title = "Cars")
}

