% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/view_duplicated.R
\name{view_duplicated}
\alias{view_duplicated}
\title{Show Duplicated Rows}
\usage{
view_duplicated(data, ...)
}
\arguments{
\item{data}{a data frame}

\item{...}{columns to search for duplicated values; defaults to all arguments}
}
\value{
a data table with duplicated rows by given columns
}
\description{
Finds all duplicated rows by given columns. Equivalent to doing data[duplicated(data) | duplicated(data, fromLast = TRUE), ]
}
\examples{
d <- data.frame(x = rep(1:3, each=4), y = rep(1:4, each=3), z = rep(1:2, 6), a = rep(1:6, 2))
view_duplicated(data = d, y, z)
view_duplicated(d, a, z)
view_duplicated(d)
}
\seealso{
\code{\link{remove_duplicated}} to remove duplicates
}

