% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_cases.R
\name{complete_cases}
\alias{complete_cases}
\title{Removes NA Observations By Column}
\usage{
complete_cases(data, ...)
}
\arguments{
\item{data}{a data frame}

\item{...}{columns to remove NA obs, defaults to all arguments; can also apply dplyr special functions}
}
\value{
a data frame
}
\description{
Deprecated. Use tidyr::drop_na.
}
