% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{diagnostic_plots}
\alias{diagnostic_plots}
\title{Diagnostic Plots for lm}
\usage{
diagnostic_plots(mod)
}
\arguments{
\item{mod}{a lm model}
}
\value{
a list of diagnostic plots
}
\description{
Generates the diagnostic plots in ggplot2, similar to the ones that are generated via the plot command. Currently available for lm models only
}

