% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_iterative_list.R
\name{is_iterative_list}
\alias{is_iterative_list}
\title{Iterative List Check}
\usage{
is_iterative_list(l)
}
\arguments{
\item{l}{a list}
}
\value{
logical
}
\description{
An iterative list is one which is generated from an lapply loop.
The elements of l does not need to be named. The elements of l
should be named objects. Each element within l should be an object 
with the same names. No element in l should be null, but it is ok 
if an element of l contains a null object (as long as it is named).
}
\examples{
x <- purrr::rerun(5, x = list(a = runif(1), b = rnorm(15)), y1 = runif(12), i = if(sample(c(TRUE, FALSE), 1)) head(iris) else NULL)
is_iterative_list(x)
}
