% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nhuyhoa.R
\name{nhuyhoa}
\alias{nhuyhoa}
\alias{nhuyhoa_df_print}
\title{Functions for nhuyhoa jekyll blog}
\usage{
nhuyhoa()

nhuyhoa_df_print(df, head = 5)
}
\description{
Functions to use when running jekyll blog
}

