% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicated.R
\name{remove_duplicated}
\alias{remove_duplicated}
\title{Removes Duplicated Rows}
\usage{
remove_duplicated(data, ..., opt_delete = "from first", opt_summary = TRUE)
}
\arguments{
\item{data}{a data frame}

\item{...}{columns to search for duplicated values; defaults to all arguments}

\item{opt_delete}{string, options for deletion; takes the values "from first", "from last", or "all". "all" deletes all duplicated entries, "from first" keeps the first duplicate and deletes the rest, "from last" keeps the last duplicate and deletes the rest. Default is "from first".}

\item{opt_summary}{boolean, whether to print drop summaries}
}
\value{
Returns a data table with duplicated rows given by columns removed
}
\description{
Removes duplicated data by given columns
}
\examples{
remove_duplicated(iris, Species, opt_delete = "from last")

}
\seealso{
\code{\link{view_duplicated}} to look at duplicates (but not remove)
}

