% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{shortcuts}
\alias{shortcuts}
\alias{blank_slate}
\alias{install_jn}
\alias{lib}
\title{Shortcut Functions}
\usage{
blank_slate()

install_jn()

lib(...)
}
\description{
Things that I commonly use in convenient functions
}
\details{
blank_slate() clears the environment and console \cr

install_jn() installs the jn.general package \cr

lib() imports libraries, see scriptR::lib() \cr
}
