% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{shortcuts}
\alias{\%p\%}
\alias{blank_slate}
\alias{install_jn}
\alias{lib}
\alias{shortcuts}
\title{Shortcut Functions}
\usage{
blank_slate()

install_jn()

lib(...)

x \%p\% y
}
\description{
Things that I commonly use in convenient functions
}
\details{
blank_slate() clears the environment and console \cr

install_jn() installs the jn.general package \cr

percent-p-percent is shortcut for paste0 \cr

lib() imports libraries depending on args: \cr
 data: (data wrangling) magrittr, stringr, plyr, dplyr, tidyr, reshape2, data.table \cr
 viz: (data visualization) knitr, ggplot2, grid, gridExtra, GGally, gtable \cr
 model: (modeling) broom, lme4, glmnet, caret \cr
 develop: (development) devtools, microbenchmark \cr

nhuyhoa() generates nhuyhoa jekyll blog
}

