% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_be.R
\name{to_be}
\alias{to_be}
\title{To Be or Not To Be}
\usage{
to_be(x, f, ...)
}
\arguments{
\item{x}{an object to be subsetted}

\item{f}{a subsetting function (subset, dplyr::slice, dplyr::filter, dplyr::select, etc) that applies to only one dimension of the data frame}

\item{...}{additional arguments to subsetting function f}
}
\value{
list \cr
Element "to_be" contains the subsetted object \cr
Element "not_to_be" contains the opposite of the subsetted object (anti-subset) \cr
}
\description{
Obtains the subset and the opposite of the subset (anti-subset) of data frames.
}
\examples{
to_be(mtcars, dplyr::slice, 1:5)
to_be(mtcars, subset, cyl != 4)
to_be(mtcars, dplyr::filter, cyl != 4, mpg > 20)
to_be(mtcars, dplyr::select, -one_of(c("gear", "vs")))
to_be(iris, view_duplicated)

}
