% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_duplicated.R
\name{view_duplicated}
\alias{view_duplicated}
\title{Show Duplicated Rows}
\usage{
view_duplicated(data, ...)
}
\arguments{
\item{data}{a data frame}

\item{...}{columns to search for duplicated values, defaults to all arguments; can also apply dplyr special functions}
}
\value{
a data frame with duplicated rows by given columns
}
\description{
Finds all duplicated rows by given columns. Equivalent to doing data[duplicated(data) | duplicated(data, fromLast = TRUE), ]
}
\examples{
d <- data.frame(x1 = rep(1:3, each=4), x2 = rep(1:4, each=3), z = rep(1:2, 6), a = rep(1:6, 2))
view_duplicated(data = d, x2, z)
view_duplicated(d, a, z)
view_duplicated(d)
view_duplicated(d, starts_with("x"))

}
\seealso{
\code{\link{remove_duplicated}} to remove duplicates
}
